chore: move `Tomcat` resource to package `tomcat/resoruce`
