+++ sample-operators/tomcat-operator/src/main/java/io/javaoperatorsdk/operator/sample/tomcat/resource/TomcatStatus.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.sample;
+package io.javaoperatorsdk.operator.sample.tomcat.resource;
 
 public class TomcatStatus {
